BREAKING: Add metadata container for the KafkaConnection
Motivation

We constantly breaking API for the Kubernetes objects.
This is not very good for already existing users using the objects.
When we merge to the OpenShift console we will be locked with the API changes.

This change adds all extra (mostly quarkus specific values from managed kafka to the metadata container). Those values are needed but if quarkus team will want use to add anything or rename things we will not break OpenShift console integration or bump version of the CR that will make this integration redundant.