fix: Add ability to preserve custom fields for status of every object.
Currently there is bug in the java operator that prevents from updating resources. However we can technically workaround it by adding some temporary fields to status. We do not want them as part of the API and support it in the future but we really need it to support end to end plugin workflow